corrected copy-paste typos in the multigrid structure (coarse_solver_maxiter type is changed to int)

+++ include/quda.h
@@ -452,8 +452,8 @@ extern "C" {
     /** Tolerance for the solver that wraps around the coarse grid correction and smoother */
     double coarse_solver_tol[QUDA_MAX_MG_LEVEL];
 
-    /** Tolerance for the solver that wraps around the coarse grid correction and smoother */
-    double coarse_solver_maxiter[QUDA_MAX_MG_LEVEL];
+    /** Maximum number of iterations for the solver that wraps around the coarse grid correction and smoother */
+    int coarse_solver_maxiter[QUDA_MAX_MG_LEVEL];
 
     /** Smoother to use on each level */
     QudaInverterType smoother[QUDA_MAX_MG_LEVEL];
